Add ROIs matching step to parcellation stage.
Fix such that the assigned ROIs in the parcellationFile matches the numbers of regions in the lookup table provided with the template. The ROIs can mismatch in case the order of the color lookup table of a template does not match the lookup table used to create the GCS file or the lookup table skips one region number.