Improve backwards compatibility reconstruction diffusion peaks
Part of CATO v3.1, including fixes that make the diffusion reconstruction step more in line with CATO v2.6. Non-linear LM solver of the DTI tensor is updated. The iRESTORE algorithm is updated and also better commented. The CSD algorithm is improved (the threshold based on tau is now forced to be positive) and will provide also a best fit if the algorithm does not converge (in line with v2.6).