Add limitation on number of used computation threads
MATLAB can utilize multiple threads in some functions. At default, the maximum number of computational threads is set to the number of physical cores. The new parameter general.maxNumberCompThreads limits the number of threads which is useful when running multiple instances of CATO in parallel on clusters. If the parameter is set to 0, CATO will use the automatic value chosen by MATLAB