Implement alternative for getNiftiProperties
getNiftiProperties provides a fast overview of the header of a nifti file using FSL's mri_info. This dependency causes errors (e.g. when FSL is not installed correctly) and complicates continuous testing. In this commit, I replace getNiftiProperties with a faster implementation of load_nifti_fmri and with a small block of code to get the orientation (LAS/RAS/etc) from the vox2ras property.