Alternative implementation for getNiftiProperties (#5)
Replace the function getNiftiProperties with code that is not dependent on FSL's mri_info. The function getNiftiProperties provided a fast overview of the header of a nifti file using FSL's mri_info. However, sometimes this dependency causes errors (e.g. when FSL is not installed correctly) and it complicates continuous integration. This branch replaces getNiftiProperties with code that uses load_nifti and an updated (faster) implementation of load_nifti_fmri.