Add test for bandpass filter step
This commit adds the test "reconstruction_functional_network_bandpass_filter" to the test suite that asserts the correct filtering of low and high noise in the bandpass filter step. The signal in the simulated rs-fMRI time series was changed from correlated random vectors from a multivariate normal distribution to a sinusoid signals at 0.05Hz.

Note, the test suite is still in development and as such the code in this commit is not yet as structured/clean as desired.