Add constant regressors to model
The test "reconstruction_functional_network_bandpass_filter" failed after the test suite was updated. This was due to a constant regressor missing from the model. The impact of this bug is most likely very small in practice as the average white matter signal regressor explains most of the "constant" signal.