Fix bandpass filter (#28)
This commit fixes the fMRI bandpass filter artefacts at the beginning and end of the filtered time series and closes Issue #27. The code changes improve the foltering process by removing the matching of the initial and final timepoints of the unfiltered signals (as done by MATLAB's `filtfilt` function), as these are often outliers from the perspective of the filtered signal. Additionally, the signal is padded using the time-mirrored timeseries instead of the inverted time-mirrored timeseries (as done by MATLAB's `filtfilt` function).

* Manually implement bandpass filter

Removes the matching of the initial and final timepoints of the unfiltered signals, as these are often outliers from the perspective of the filtered signal;
Pads the signal using the the time-mirrored timeseries, not the inverted time-mirrored timeseries.

* Add test for bandpass filter step

This commit adds the test "reconstruction_functional_network_bandpass_filter" to the test suite that asserts the correct filtering of low and high noise in the bandpass filter step. The signal in the simulated rs-fMRI time series was changed from correlated random vectors from a multivariate normal distribution to a sinusoid signals at 0.05Hz.

Note, the test suite is still in development and as such the code in this commit is not yet as structured/clean as desired.

* Add constant regressors to model

The test "reconstruction_functional_network_bandpass_filter" failed after the test suite was updated. This was due to a constant regressor missing from the model. The impact of this bug is most likely very small in practice as the average white matter signal regressor explains most of the "constant" signal.

---------

Co-authored-by: Siemon de Lange <SiemondeLange@users.noreply.github.com>