Add setupPythonScript parameter
Instead of specifying the location of the python binary in the `pythonInterpreter` parameter, users can now specify a start-up script with the `pythonInterpreter` parameter that can activate their desired environment. This change has the benefit that the COMMIT script also works when no startup script is provided and that it is more in line with the usual usage of activating Anaconda environments from the command line.