Handling synb0File
Issue
Trying to incorporate synb0-DISCO in the CATO workflow it was not compatible. When setting the reverseB0Scan to the synb0 disco, it caused errors due to autogenerated index and acqp, and merging of dwi files.

Solution
A new preprocessing script was added. Structural_preprocessing.m now recognizes the synb0File setting in configParams. When detected the following steps are taken:
- check dimensions of input data, these should be even due to topup config
- uses indexFile and acqpFile as INPUT. The echo time of the synthesized b0 image should be zero
- if indexFile and acqpFile are not located in outputDir a copy is made to keep track of used files.
- errors following checks were added